+++ core/meterer/meterer_test.go
@@ -495,7 +495,7 @@ func TestMeterer_symbolsCharged(t *testing.T) {
 func createPaymentHeader(timestamp int64, cumulativePayment *big.Int, accountID gethcommon.Address) *core.PaymentMetadata {
 	return &core.PaymentMetadata{
 		AccountID:         accountID.Hex(),
-		Timestamp:         uint64(timestamp),
+		Timestamp:         timestamp,
 		CumulativePayment: cumulativePayment,
 	}
 }
